+++ trixnity-client/src/commonTest/kotlin/de/connect2x/trixnity/client/mocks/MediaServiceMock.kt
@@ -15,6 +15,15 @@ class MediaServiceMock : MediaService {
     override suspend fun getMedia(
         uri: String,
         progress: MutableStateFlow<FileTransferProgress?>?,
+        saveToCache: Boolean,
+    ): Result<PlatformMedia> {
+        throw NotImplementedError()
+    }
+
+    override suspend fun getMedia(
+        uri: String,
+        maxSize: Long?,
+        progress: MutableStateFlow<FileTransferProgress?>?,
         saveToCache: Boolean
     ): Result<PlatformMedia> {
         throw NotImplementedError()
@@ -22,6 +31,15 @@ class MediaServiceMock : MediaService {
 
     override suspend fun getEncryptedMedia(
         encryptedFile: EncryptedFile,
+        progress: MutableStateFlow<FileTransferProgress?>?,
+        saveToCache: Boolean,
+    ): Result<PlatformMedia> {
+        throw NotImplementedError()
+    }
+
+    override suspend fun getEncryptedMedia(
+        encryptedFile: EncryptedFile,
+        maxSize: Long?,
         progress: MutableStateFlow<FileTransferProgress?>?,
         saveToCache: Boolean
     ): Result<PlatformMedia> {
@@ -32,6 +50,19 @@ class MediaServiceMock : MediaService {
         uri: String,
         width: Long,
         height: Long,
+        method: ThumbnailResizingMethod,
+        animated: Boolean,
+        progress: MutableStateFlow<FileTransferProgress?>?,
+        saveToCache: Boolean,
+    ): Result<PlatformMedia> {
+        throw NotImplementedError()
+    }
+
+    override suspend fun getThumbnail(
+        uri: String,
+        width: Long,
+        height: Long,
+        maxSize: Long?,
         method: ThumbnailResizingMethod,
         animated: Boolean,
         progress: MutableStateFlow<FileTransferProgress?>?,