+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added automatic cleanup of media, when canceling outbox message
 - Added UriExtractor to `EventContentMediaMappings` (previously `OutboxMessageMediaUploaderMappings`)
+- download size limit for media service methods
 
 ### Changed
 
@@ -26,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SignedPersistentDataUnit<T>` type alias
 - `SignedPersistentStateDataUnit<T>` type alias
 - `SignedPersistentMessageDataUnit<T>` type alias
+- `MediaService`: `getMedia`, `getEncryptedMedia` and `getThumbnail` without required `maxSize` parameter
 
 ### Removed
 
@@ -37,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed c2x-conventions update requiring GitLab token
 
 ### Security
+
+- prevent resource exhaustion risk by adding download limit parameters to `MediaService`
 
 ## 5.6.0